Fix initial inode creation
The initial inode creation was going fine, but the inode
never got written out again.  That means that all writes
produced zero-length files, although there *were* block
written.

The issue was that the inode wasn't getting marked dirty.
I think this had to do with the writeout code assuming
that all inodes for which it could not get a lock would
eventually get properly written out, and it would still
remove those from the queue.

But, the lock is taken for other things, and there are
probably still races there anyway.

So, wirteout threads will *always* either write out the
object or place it back in the queue.  Extra writeouts
should be avoided by the writeout code checking __dirty
before actually doing writeouts.