Make more tolerant of small writes
It used to be that we had a 50-operation limit in the queue.
That meant that _any_ 50 writes would block until the queue
got shrunk.

Now, we at least ensure that those 50 writes are to 50 unique
objects.  That means that 1000 small writes to an object will
only end up counting for one of the 50 spots in that queue.
Well, actually 2 spots (you need one for the dirty inode).